fix compiler warnings for -Wall and MSVC
Mostly int/size_t mismatches.

+++ ImFileDialog.h
@@ -108,7 +108,7 @@ namespace ifd {
 
 		std::string m_filter;
 		std::vector<std::vector<std::string>> m_filterExtensions;
-		int m_filterSelection;
+		size_t m_filterSelection;
 		void m_parseFilter(const std::string& filter);
 
 		std::vector<int> m_iconIndices;
